Responding to PR comments.
