+++ pypeit/core/pca.py
@@ -268,6 +268,7 @@ def get_pc(data, k, tol=0.0, maxiter=20, nofix=False, noortho=False):
         hidden = np.dot(eigv.T, data)
     return eigv, hidden
 
+## TODO I think all of this code below is actually deprecated.
 ################################################
 # 2D Image PCA
 
@@ -548,95 +549,4 @@ def pca_plot(setup, inpar, ofit, prefix, maxp=25, pcadesc="", addOne=True,
     return
 
 
-def pca_arctilt(slf, tiltang, centval, tilts, maxp=25, maskval=-999999.9):
-    """ Generate a QA plot for the blaze function fit to each slit
-
-    Parameters
-    ----------
-    slf : class
-      Science Exposure class
-    tiltang : ndarray
-      (m x n) 2D array containing the measured tilts (m) for each slit (n)
-    centval : ndarray
-      (m x n) 2D array containing the pixel location (in the spectral direction) of the measured tilts (m) for each slit (n)
-    tilts : ndarray
-      (m x n) 2D array containing the model tilts (m) for each slit (n)
-    maxp : int, (optional)
-      Maximum number of panels per page
-    maskval : float, (optional)
-      Value used in arrays to indicate a masked value
-    """
-    plt.rcdefaults()
-    plt.rcParams['font.family']= 'times new roman'
-    
-    method = inspect.stack()[0][3]
-    outroot = arqa.set_qa_filename(slf.setup, method)
-    #
-    npc = tiltang.shape[1]
-    pages, npp = arqa.get_dimen(npc, maxp=maxp)
-    x0 = np.arange(tilts.shape[0])
-    # First calculate the min and max values for the plotting axes, to make sure they are all the same
-    w = np.where(tiltang != maskval)
-    medv = np.median(tiltang[w])
-    madv = 1.4826 * np.median(np.abs(medv - tiltang[w]))
-    ymin, ymax = medv - 3.0 * madv, medv + 3.0 * madv
-    ymin = min(ymin, np.min(tilts))
-    ymax = max(ymax, np.max(tilts))
-    # Check that ymin and ymax are set, if not, return without plotting
-    if ymin is None or ymax is None:
-        msgs.warn("Arc tilt fits were not plotted")
-        return
-    # Generate the plots
-    ndone = 0
-    for i in range(len(pages)):
-        f, axes = plt.subplots(pages[i][1], pages[i][0])
-        ipx, ipy = 0, 0
-        for j in range(npp[i]):
-            if pages[i][1] == 1:
-                ind = (ipx,)
-            elif pages[i][0] == 1:
-                ind = (ipy,)
-            else:
-                ind = (ipy, ipx)
-            w = np.where(tiltang[:, ndone] != maskval)
-            if np.size(w[0]) != 0:
-                axes[ind].plot(centval[:, ndone][w], tiltang[:, ndone][w], 'bx')
-                axes[ind].plot(x0, tilts[:, ndone], 'r-')
-            axes[ind].axis([0, tilts.shape[0]-1, ymin, ymax])
-            axes[ind].set_title("Slit {0:d}".format(1+ndone))
-            axes[ind].tick_params(labelsize=8)
-            ipx += 1
-            if ipx == pages[i][0]:
-                ipx = 0
-                ipy += 1
-            ndone += 1
-        # Delete the unnecessary axes
-        for j in range(npp[i], axes.size):
-            if pages[i][1] == 1:
-                ind = (ipx,)
-            elif pages[i][0] == 1:
-                ind = (ipy,)
-            else:
-                ind = (ipy, ipx)
-            f.delaxes(axes[ind])
-            ipx += 1
-            if ipx == pages[i][0]:
-                ipx = 0
-                ipy += 1
-        # Save the figure
-        if pages[i][1] == 1 or pages[i][0] == 1:
-            ypngsiz = 11.0/axes.size
-        else:
-            ypngsiz = 11.0*axes.shape[0]/axes.shape[1]
-        f.set_size_inches(11.0, ypngsiz)
-        f.tight_layout()
-        outfile = outroot+'{:02d}.png'.format(i)
-        plt.savefig(outfile, dpi=200)
-        plt.close()
-        #f.clf()
-        #del f
-
-    plt.rcdefaults()
-
-    return
